refactor(arch-hexagonal) : refactor for patterns arch hexagonal
