+++ src/main/java/br/com/branas/glstore/application/services/DiscountCouponService.java
@@ -1,6 +1,6 @@
-package br.com.branas.glstore.domain.services;
+package br.com.branas.glstore.application.services;
 
-import br.com.branas.glstore.domain.entities.DiscountCoupon;
+import br.com.branas.glstore.application.entities.DiscountCoupon;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
