+++ src/main/java/br/com/branas/glstore/application/services/ProductService.java
@@ -1,6 +1,6 @@
-package br.com.branas.glstore.domain.services;
+package br.com.branas.glstore.application.services;
 
-import br.com.branas.glstore.domain.entities.Product;
+import br.com.branas.glstore.application.entities.Product;
 import org.springframework.stereotype.Service;
 
 import java.util.Arrays;
