+++ src/main/java/br/com/branas/glstore/infrastructure/repositories/OrderRepository.java
@@ -1,6 +1,6 @@
 package br.com.branas.glstore.infrastructure.repositories;
 
-import br.com.branas.glstore.domain.entities.Order;
+import br.com.branas.glstore.application.entities.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
